Rename to WmscCapabilitiesParser
