+++ BruTile/Wms/WmsCapabilities.cs
@@ -10,9 +10,6 @@ using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.Linq;
 
-// Disabled warning for obsolete WebRequest because this is not tested in our code so we can not
-// validate an alternative solution. We do not use this ourselves so do not want to spend time on it.
-
 namespace BruTile.Wms;
 
 public class WmsCapabilities
